content/media/test/test_timeupdate_small_files.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=495319
-->

<head>
  <title>Bug 495319 - playing back small audio files should fire timeupdate</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=495319">Mozilla Bug 495319</a>
<pre id="test">
<script class="testbody" type="text/javascript">

var manager = new MediaTestManager;

function timeupdate(e) {
  var v = e.target;
  v._timeupdateCount++;
  ok(v.gotEnded == 0, v._name + " - shouldn't get timeupdate after ended");
}

function ended(e) {
  var v = e.target;
  ++v.gotEnded;
  ok(v._timeupdateCount > 0, v._name + " - should see at least one timeupdate: " + v.currentTime);
  v._finished = true;
  // Finish the test after 500ms. We shouldn't receive any timeupdate events
  // after the ended event, so this gives time for any pending timeupdate events
  // to fire so we can ensure we don't regress behaviour.
  setTimeout(
    function() {
      // Remove the event listeners before removing the video from the document.
      // We should receive a timeupdate and pause event when we remove the element
      // from the document (as the element is specified to behave as if pause() was
      // invoked when it's removed from a document), and we don't want those
      // confusing the test results.
      v.removeEventListener("ended", ended, false);
      v.removeEventListener("timeupdate", timeupdate, false);
      v.removeEventListener("timeupdate", logEvent, false);
      v.removeEventListener("pause", logEvent, false);
      v.removeEventListener("stalled", logEvent, false);
      v.parentNode.removeChild(v);
      manager.finished(v.token);
    },
    500);
}

var eventsToLog = ["play", "canplay", "canplaythrough", "loadstart", "loadedmetadata",
  "loadeddata", "playing", "progress", "timeupdate", "ended", "suspend", "error", "stalled", "emptied", "abort",
  "waiting", "pause"];
function logEvent(event) {
  if (event.target.gotEnded > (event.type == "ended" ? 1 : 0)) {
    ok(false, event.target.currentSrc + " got unexpected " + event.type + " after ended");
  } else {
    info(event.target.currentSrc + " got " + event.type);
  }
}

function startTest(test, token) {
  var type = /^video/.test(test.type) ? "video" : "audio";
  var v = document.createElement(type);
  v.token = token;
  manager.started(token);
  v.src = test.name;
  v._name = test.name;
  v._timeupdateCount = 0;
  v._finished = false;
  v.gotEnded = 0;
  v.autoplay = true;
  v.addEventListener("ended", ended, false);
  v.addEventListener("timeupdate", timeupdate, false);
  for (var i = 0; i < eventsToLog.length; ++i) {
    v.addEventListener(eventsToLog[i], logEvent, false);
  }
  document.body.appendChild(v);
}

manager.runTests(gSmallTests, startTest);

</script>
</pre>
</body>
</html>
